% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bostypologie_btyp_calculateAffinity.R
\name{btyp_calculateAffinity}
\alias{btyp_calculateAffinity}
\title{Calculate affinity-score with a certain forest type}
\usage{
btyp_calculateAffinity(dat, ref, correct = TRUE, typology = "T1")
}
\arguments{
\item{dat}{Opnamedata}

\item{ref}{Referentiedata}

\item{correct}{Corrigeer op basis van vegetatiebedekking}

\item{typology}{Welke typologie wil je gebruiken
\itemize{
\item{T1}{Categorisatie in 10 hoofdtypes}
\item{T2}{Categorisatie in 30 subtypes}
\item{T3}{Categorisatie in 39 detailtypes}}}
}
\description{
This function is a wrapper that calculates for every "vegetatieopname" the similarity score using the function btyp_calculateSimilarity
}
