% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bostypologie_btyp_calculateSimilarity.R
\name{btyp_calculateSimilarity}
\alias{btyp_calculateSimilarity}
\title{Calculate similarities}
\usage{
btyp_calculateSimilarity(dat, ref, correct, typology = "T1")
}
\arguments{
\item{dat}{Vegetation data only containing 1 "vegetatieopname"}

\item{ref}{Reference data to which the similarity is calculated}

\item{correct}{Correct for vegetation density or not}

\item{typology}{Which typology score to use (now T1 (10 types), T2 (30 subtypes), T3 (39 detail types) are allowed)}
}
\description{
The central routine to calculate similarities. Normally called from a do (dplyr) or ddply (plyr) loop like btyp_calculateAffinity
}
