% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bostypologie_btyp_calibrateForestTypology.R
\name{btyp_calibrateForestTypology}
\alias{btyp_calibrateForestTypology}
\title{Title}
\usage{
btyp_calibrateForestTypology(obj, typology = "T1")
}
\arguments{
\item{typology}{}
}
\description{
Title
}
