% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bostypologie_btyp_makeCircle.R
\name{btyp_makeCircle}
\alias{btyp_makeCircle}
\title{Create xy coordinates of a circle given the radius and resolution}
\usage{
btyp_makeCircle(radius, num)
}
\arguments{
\item{radius}{radius of circle}

\item{num}{resolution of circle. Amount of points of which a half-circle consists}
}
\value{
dataset met cirkelcoordinaten
}
\description{
Create xy coordinates of a circle given the radius and resolution
}
\examples{
cko <- btyp_makeCircle(radius = 3, num = 100)
par(pty="s")
plot(cko$x, cko$y, type="l", pty = "s")
}
