% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bostypologie_btyp_plotAffinity.R
\name{btyp_plotAffinity}
\alias{btyp_plotAffinity}
\title{Title}
\usage{
btyp_plotAffinity(obj, size = 2)
}
\arguments{
\item{obj}{object waarop berekeningen gebeuren}

\item{size}{dikte van de lijn}
}
\value{
ggplot object
}
\description{
Title
}
