% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bostypologie_btyp_prepareSampleData.R
\name{btyp_prepareSampleData}
\alias{btyp_prepareSampleData}
\title{Title}
\usage{
btyp_prepareSampleData(opname, scale, conversiontable, specieslist,
  lang_sp = "NL", namecolumn = "SOORT", keycolumn = "OPNAMECODE",
  coveragecolumn = "COVERAGE_CODE")
}
\arguments{
\item{coveragecolumn}{}
}
\description{
Title
}
