\name{inbobostypologie}
\alias{inbobostypologie-package}
\alias{inbobostypologie}
\docType{package}
\title{
\packageTitle{inbobostypologie}
}
\description{
\packageDescription{inbobostypologie}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{inbobostypologie}
\packageIndices{inbobostypologie}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{inbobostypologie}

Maintainer: \packageMaintainer{inbobostypologie}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
